PC-lint for schedule commands

+++ HimalayaMain/Master/Components/Scheduler/Commands/Include/CmdALControlValve.h
@@ -139,7 +139,7 @@ public:
      *  \return from GetResult
      */
     /****************************************************************************/
-	QString GetStrResult()const{ return QString("%1").arg(m_result);}
+    QString GetStrResult()const{ return QString("%1").arg((qint32)m_result);}
 
 private:
     CmdALControlValve();                                              ///< Not implemented.
